Rename sanksjon types
