+++ apps/etterlatte-saksbehandling-ui/client/src/shared/types/sanksjon.ts
